Update dependencies
The main motiviation is to get `crossbeam-channel` at at least 0.5.1 as
this appears to fix a memory like[1] which is probably the root-cause of
our own memory issues[2].

This also included some changes to the LSP type-system as a few more
changes in the LSP protocol are now incorporated, such as a
clarification that numbers are usually unsigned 32bit rather than
64bit[3].

[1] https://github.com/crossbeam-rs/crossbeam/blob/crossbeam-channel-0.5.1/crossbeam-channel/CHANGELOG.md#version-051
[2] https://github.com/nix-community/rnix-lsp/issues/33
[3] https://github.com/gluon-lang/lsp-types/commit/f654090c3cfe291c07433bc155dadb0da3389197

+++ src/utils.rs
@@ -65,10 +65,10 @@ pub fn lookup_pos(code: &str, pos: Position) -> Option<usize> {
 pub fn offset_to_pos(code: &str, offset: usize) -> Position {
     let start_of_line = code[..offset].rfind('\n').map_or(0, |n| n + 1);
     Position {
-        line: code[..start_of_line].chars().filter(|&c| c == '\n').count() as u64,
+        line: code[..start_of_line].chars().filter(|&c| c == '\n').count() as u32,
         character: code[start_of_line..offset]
             .chars()
-            .map(|c| c.len_utf16() as u64)
+            .map(|c| c.len_utf16() as u32)
             .sum(),
     }
 }
